Add Flightplan logic to start/restart app
Rather than reload all Node apps managed with PM2 on the server,
I made sure to start the app with a unique name, and then I can
check for that name on deploy. If an app with that name is running,
I just reboot it. Otherwise, I start it fresh.