Add turnPoints calculations + presentation
One vital feature that was missing was the ability to see what a
tentative turn is worth before submitting. I wrote some logic in the
game reducer that figures out how many points its worth, and what the
last tile in the placed word is.

Then, I added some logic in the BoardSquare component to display a
<div> containing the turnPoints value if it's present on the tile.

A bit of CSS to round it out, and we now have a Words With Friends
style point tally.