Add passTurn on the server. Moved `isMyTurn` to server.
I implemented the GameSchema instance method 'passTurn' to handle
turns that didn't have a submitted word attached. As part of the
test suite, I wanted to make sure that a user couldn't pass their
competitor's turn, by calling `passTurn` when it wasn't their turn.

In doing so, I realized that there wasn't any easy way to tell, on
the server, whose turn it was. This led to the realization that this
is really a piece of data that should live on the server, and be sent
to the client. As a result, game.selector is now much cleaner, as it
just needs to display a property from the server instead of computing
it.