Improve board-to-rack tile drop behaviour
Finally, I have suitable board-to-rack tile dropping. When a tile
from the board is dropped in position 2, it takes that tile's
position and shifts all subsequent tiles down 1 spot.

The logic could be better: It doesn't support rack-to-rack dropping
in the same way, so that could be confusing. I need to straighten
out the X position stuff.