Start building Swap components
Started the groundwork for building out TileSwapContainer and
TileSwapBucket. Still need to implement the DnD and the CSS, but
the basic skeleton is complete.

Also refactored all SidePanel sub-components to not include SidePanel
in their name (eg. SidePanelPlayers -> Players). This way the
components aren't glued so heavily to their context; I could use the
Players component elsewhere if I need to.