Refactor Swap to be a simple array of tiles.
Before swap, there were two possible locations: `rack` and `board`.
Both were properties on the 'game' slice of the reducer, and both
were simple arrays of tiles.

Swap was differnt; it was an object that held a boolean and an array
of tiles.

Because the structure was different, a lot of my convenience methods
for placing tiles or identifying tile location weren't compatible.

By turning `swap` into a plain array of tiles, and moving the boolean
to a property on the 'game' slice, I greatly simplified lots of areas
in the code =)