Implemented pick_tile/place_tile sounds
Huzzah, there is sound. Got a tiny lib file for interfacing with
Howler to play audio snippets, and called it from a few different
components when the action occurs.

This will likely need to change, since not all sound effects are
the direct result of the user's action (for example, the "it's your
turn" sound effect can happen while the user is nowhere near the
computer).

To avoid having logic spread across a bunch of different types of
files, I'm thinking I'll move it to the action creators. This commit,
then, is just for reference.