Fix bug with gamesList reducer
Formerly, every time the interface would request the games list,
the server would respond with the games and the reducer would .concat
these games in, even if it already had them. THis was leading to the
reducer having duplicate copies of the same game.

Turns out the fix was super simple: Just using an OrderedSet instead
of a List, in immutablejs. This way, duplicate additions are
automatically ignored.