Finish actions on gameList
The gamesList has an 'actions' column, and it was pretty buggy.
Now it changes depending on the status of the game, whether the
user is authenticated, and whether the authenticated user is already
playing in a specific game.

This also fixes the issue, on the client, of a user joining the same
game multiple times. A server check should be added for redundancy.