Start working on results display
Made a modal that shows up when the game is completed, that displays
a "you win"/"You lose" message. Changed the selector to be a boolean,
and added the game status to the selector.

Will be used both for
showing/hiding the results screen but also for some game logic (eg.
no moving while the game is "waiting").