Add end-game condition check
Checked, when submitting a word, if it was enough to end the game.
If so, set the state to 'completed' and stored the winnerUserId on
the game.

Updated the client selector to use this new winnerUserId field,
instead of trying to derive it on the client. Once surrendering /
abandoning games becomes a thing, the winner may not always be the
user with the highest points ;)