Add and style 'waiting' indicator.
To stop a player from being able to play while the game was in
'waiting' mode, I hid the TileRack and Controls. I also, though,
needed some sort of visual indicator that the game was waiting for
players.

I added a div with some subtle animations so it felt like the page
wasn't static/unchanging.