Tweak CSS, add button to leave completed games.
As a band-aid before allowing modals to be dismissed, added a
button to return to the games list once a game finishes and the
win/lose modal pops up.

Some misc styling and cleanup as well. Rearranged the constants to
be alphabetical, as I often try to import them in this order.